made webapp. Fixed config.py
