+++ requirements.txt
@@ -2,3 +2,4 @@
 Flask==2.1.2
 requests==2.27.1
 tinkoff-investments==0.2.0b29
+jupyter==1.0.0
