+++ webapp/config.py
@@ -1,0 +1,3 @@
+currencies_API_KEY = "PHYlozjcZ5z6aDwXStIkjSYq40KWfRDi"
+TINKOFF_API_KEY = "t.KuJkD0UfDsdvKMYCE7iAxDFEBp9Wn46EcKUK3CVGFqnsT_EqHKxuqByOWzT-zCYBAKKJz0VYOJ1LX5fMHnIO0g" # IIS
+currencies_URL = "https://api.apilayer.com/fixer/latest"
