+++ webapp/templates/index.html
@@ -22,6 +22,7 @@
                 {% endfor %}
             </div>
         </div>
+        <div>{{ portfolio }}</div>
     </div>
 </body>
